+++ packages/observe/unix/observe_unix.ml
@@ -1,6 +1,11 @@
 module Platform = struct
   type t = unit
 
+  let terminal_style () =
+    Terminal.style
+      ~isatty:(fun () -> Unix.isatty Unix.stderr)
+      ~getenv:Sys.getenv_opt
+
   let now () = Clock.instant (Ptime_clock.now_d_ps ())
 
   let write_terminal () value =